+++ src/app/decorators/filter.ts
@@ -9,7 +9,7 @@ import { MetaData } from '../constants/MetaData';
  * Provides operations to intercept HTTP requests to a controller.
  */
 export interface IActionFilter {
-	execute(request: any, response: any, next: Function): void;
+	execute(request: any, response: any, next: Function, ...params: any[]): void | Promise<void>;
 }
 
 /**
@@ -28,10 +28,14 @@ export enum FilterPriority { LOW, MEDIUM, HIGH }
 
 export type FilterDecorator = <T extends ActionInterceptor>(
 		FilterClass: Newable<T>,
-		priority?: FilterPriority
+		priority?: FilterPriority,
+		...filterParams: any[]
 	) => Function;
 
-export type FilterArray<T extends ActionInterceptor = ActionInterceptor> = Newable<T>[];
+export type FilterArray<T extends ActionInterceptor = ActionInterceptor> = {
+	FilterClass: Newable<T>,
+	filterParams: any[]
+}[];
 export type PrioritizedFilterArray = FilterArray[];
 
 
@@ -40,10 +44,11 @@ export type PrioritizedFilterArray = FilterArray[];
  * @param {class} FilterClass Filter class whose name must end with "Filter".
  * @param {FilterPriority} priority Filters with greater priority run before ones with less priority.
  */
-export function filter<T extends ActionInterceptor>(FilterClass: Newable<T>, priority: FilterPriority = FilterPriority.MEDIUM): Function {
+export function filter<T extends ActionInterceptor>(FilterClass: Newable<T>, priority: FilterPriority = FilterPriority.MEDIUM,
+	...filterParams: any[]): Function {
 
 	return function (TargetClass: Newable<T>, key: string): Function {
-		return addFilterToTarget(FilterClass, TargetClass, key, priority);
+		return addFilterToTarget(FilterClass, TargetClass, key, priority, ...filterParams);
 	};
 }
 
@@ -56,7 +61,8 @@ export function filter<T extends ActionInterceptor>(FilterClass: Newable<T>, pri
  * @param {FilterPriority} priority Filters with greater priority run before ones with less priority.
  */
 export function addFilterToTarget<T extends ActionInterceptor>(FilterClass: Newable<T>,
-		TargetClassOrPrototype: Newable<T>, targetFunc?: string, priority: FilterPriority = FilterPriority.MEDIUM): Function {
+		TargetClassOrPrototype: Newable<T>, targetFunc?: string, priority: FilterPriority = FilterPriority.MEDIUM,
+		...filterParams: any[]): Function {
 
 	let metaKey: string,
 		isClassScope = (!targetFunc); // If `targetFunc` has value, `targetClass` is "prototype" object, otherwise it's a class.
@@ -73,7 +79,7 @@ export function addFilterToTarget<T extends ActionInterceptor>(FilterClass: Newa
 		: Reflect.getMetadata(metaKey, TargetClassOrPrototype, targetFunc);
 	filters = filters || [];
 
-	pushFilterToArray(filters, FilterClass, priority);
+	pushFilterToArray(filters, FilterClass, priority, ...filterParams);
 	Reflect.defineMetadata(metaKey, filters, TargetClassOrPrototype, targetFunc);
 	return TargetClassOrPrototype;
 }
@@ -81,9 +87,9 @@ export function addFilterToTarget<T extends ActionInterceptor>(FilterClass: Newa
 /**
  * Prepares a filter then push it to given array.
  */
-export function pushFilterToArray<T extends ActionInterceptor>(filters: PrioritizedFilterArray, FilterClass: Newable<T>, priority: FilterPriority = FilterPriority.MEDIUM): void {
+export function pushFilterToArray<T extends ActionInterceptor>(filters: PrioritizedFilterArray, FilterClass: Newable<T>, priority: FilterPriority = FilterPriority.MEDIUM, 
+		...filterParams: any[]): void {
 	Guard.assertIsTruthy(priority >= FilterPriority.LOW && priority <= FilterPriority.HIGH, 'Invalid filter priority.');
-
 
 	// `filters` is a 2-dimensioned matrix, with indexes are priority value,
 	//   values are array of Filter classes. Eg:
@@ -92,5 +98,5 @@ export function pushFilterToArray<T extends ActionInterceptor>(filters: Prioriti
 	//		3: [ FilterClass, FilterClass ]
 	// ]
 	filters[priority] = filters[priority] || [];
-	filters[priority].push(FilterClass);
+	filters[priority].push({ FilterClass, filterParams });
 }