+++ src/app/filters/ActionFilterBase.ts
@@ -1,0 +1,13 @@
+export abstract class ActionFilterBase {
+
+	protected addReadonlyProp(obj: object, prop: string, value: any): void {
+		Object.defineProperty(obj, prop, 
+			{
+				writable: false,
+				enumerable: true,
+				configurable: false,
+				value
+			}
+		);
+	}
+}