+++ src/test/decorators/authorized.spec.ts
@@ -1,5 +1,5 @@
 import * as path from 'path';
-import { describe, it } from 'mocha';
+
 import * as chai from 'chai';
 import * as spies from 'chai-spies';
 chai.use(spies);