+++ src/test/decorators/controller.spec.ts
@@ -1,6 +1,6 @@
 /// <reference types="reflect-metadata" />
 import * as path from 'path';
-import { describe, it } from 'mocha';
+
 import * as chai from 'chai';
 import * as spies from 'chai-spies';
 chai.use(spies);
@@ -9,7 +9,7 @@ import * as request from 'request-promise';
 import { injectable, DependencyContainer, serviceContext, CriticalException,
 	IConfigurationProvider, Maybe, Types as CmT, constants } from '@micro-fleet/common';
 
-import { ExpressServerAddOn, AuthAddOn, Types as T, decorators } from '../../app';
+import { ExpressServerAddOn, Types as T, decorators } from '../../app';
 
 // For typing only
 import * as Bluebird from 'bluebird';
@@ -56,7 +56,6 @@ describe('@controller()', function() {
 			serviceContext.setDependencyContainer(container);
 			container.bindConstant(CmT.DEPENDENCY_CONTAINER, container);
 			container.bind(CmT.CONFIG_PROVIDER, MockConfigurationProvider).asSingleton();
-			container.bind(T.AUTH_ADDON, AuthAddOn).asSingleton();
 			container.bind(T.WEBSERVER_ADDON, ExpressServerAddOn).asSingleton();
 
 			server = container.resolve(T.WEBSERVER_ADDON);