+++ src/test/decorators/filter.spec.ts
@@ -1,6 +1,6 @@
 /// <reference types="reflect-metadata" />
 import * as path from 'path';
-import { describe, it } from 'mocha';
+
 import * as chai from 'chai';
 import * as spies from 'chai-spies';
 chai.use(spies);